dom observer rework
